ATTENTION:
in the code you will find a path like this “E:\\documents\\Python\\files\\xauusd_database.csv” you must replace this with the absolute path of the csv file. 
Remember, the absolute path of a file is the directory + the name of the file where it is located. Remember also to put this: '\\' , when separating each directory so that postgresql interprets it well.